tried type casting

+++ 0x10-variadic_functions/1-print_numbers.c
@@ -20,10 +20,10 @@ void print_numbers(const char *separator, const unsigned int n, ...)
 	}
 
 	va_start(ptr, n);
-	for (i = 0; i < n; i++)
+	for (i = 0; i < (unsigned int)n; i++)
 	{
 		printf("%i", va_arg(ptr, int));
-		if (i < n - 1 && separator)
+		if (i < (unsigned int)n - 1 && separator)
 			printf("%s", separator);
 	}
